Added support for arrays of unknown length at the end of a struct.  In particular, you can now have:
struct Ptr<`a> {
  <`i>
  tag_t<`i> size;
  `a elts[valueof(`i)];
};

and write functions to read and write the elements.  This
required that I augment pattern matching to allow you to
extract an array reference under an existential, only when
the existential is itself of abstract type (long story that
only Dan will understand.)

While I was at it, I ended up fixing some broken pieces
in pattern matching having to do with binding a variable
to an array.  Now, as expected, the variable gets promoted
to a pointer type.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2638 581db07f-6e62-4d40-9733-d432c4eebb07