* We now always use the generated cyc_setjmp instead of including <setjmp.h> in the output.  This avoids dependencies that might conflict as was the case with sigaction defined in libc.cys.  In particular, libc.cys defined:
  extern "C" int sigaction(int, const struct sigaction *,
                           struct sigaction *);

This gets translated to:

  int sigaction(int, const struct Cyc_sigaction *,
                struct Cyc_sigaction *);

However, on at least Cygwin, the ouput C code also includes
#include <setjmp.h> which in turn includes <signal.h>.  This
defines:

  int sigaction(int, const struct sigaction *, struct sigaction *);

and the C compiler doesn't like that these are incompatible types.
By getting rid of the #include <setjmp.h> and instead using
the generated cyc_setjmp, we get rid of the incompatibility.

* A little cleanup of the output to speed up gcc compilation:
not so many casts and we reduce (*e).f to e->f in most places.

* Change of syntax for the alias construct in preparation for
fixing it to get rid of the unsoundnesses.  Instead of:

  alias <`r> x = (T)e;

you should treat it more like a declaration:

  alias <`r> T x = e;

Here, `r can occur free within T, but not e.  So, Mike, you should
take all of your code that was of the old form and rewrite it.

Right now, all I'm checking is that the type of e is a subtype
of T[`U/`r], and that T is a pointer type.  This is not yet sound,
but is closer than what we had.

* Fixed bug in subtraction of thin pointers that Dan found.  When we
have e1 - e2 and e1 and e2 are both pointer types with the same
element type, you get the number of elements between e1 and e2
regardless as to whether the pointers are thin or fat.  Interestingly,
this forced me to explicitly cast some pointers to integers
in CfFlowInfo::root_cmp to get the right behavior for comparison
of variable declarations.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2643 581db07f-6e62-4d40-9733-d432c4eebb07