Moved the elimination of region partial-order constraints from the function call to the universal-elimination (i.e. type instantiation). This is complicated by the fact that we eliminate universals in two places: tcInstantiate (rarely used) and tcExpNoPromote (constantly used).
Having fixed this (it was sound before but not as useful), I fixed the
closure library (which had a wrong implementation of curry, the
correct implementation now type-checks).

May we all have advisors that read our dissertations this carefully.

Exposing a new function in tcutil.h causes the changes to irrelevant
genfiles.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2655 581db07f-6e62-4d40-9733-d432c4eebb07