Fixed bug that Bart found where we were not propagating an effective const qualifier when using x.y as an lvalue.  For instance, if x is const, then x.y = 3 should not be allowed.  This is a stop-gap fix for the overhaul on qualifiers that we really need.
Changed tcpat so that if you use an enum to match against an integer,
the span is treated as infinite.  This has the effect that we don't
give a redundant case error if you happen to write down all of the
enum constructors while matching against an integer.  For instance,
before if you had:

enum Foo { V1, V2 };

void foo(int i) {
  switch (i) {
  case V1: ...
  case V2: ...
  default: ...
  }
}

you'd get an error because the pattern-match compiler felt that the
span of the patterns was 2 instead of 2^32.  Now you don't get an
error.  However, if you write:

  switch (i) {
  case V1: ...
  case 0: ...
  default: ...
  }

you also don't get an error at the Cyclone level, because in general,
I can't evaluate enum tags down to an integer value, so I don't even
bother.  Rather, this will be caught as an error at the C level.

Still to do: we need to check that when you cast a value to an enum,
that the value is one of the cases in the enum.  Alternatively, we
could always make the span of enum's infinite, but then we'd have to
add a lot of cases in the compiler to cover situations where we
pattern match on enums.  Actually, this may be the right thing to
do since we can't really check for non-overlapping instances at
compile time, since we can't always evaluate the enum tag expressions.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2668 581db07f-6e62-4d40-9733-d432c4eebb07