A new file tests/fail1.cyc has a bunch of simple type errors in it that are meant to exercise all of the places where we call terr(...).  The target failtest in tests/Makefile should run the compiler on this file, collect the error messages, and compare them to tests/fail1.out.  You should see one error message but not two.  If you see the second error message, it indicates that something has changed and we need to figure out whether we've inroduced a bug or need to update fail1.out.
The tests uncovered a lot of minor bugs which I fixed.

* Changed the way Tvars and Evars are pretty-printed and dumped
to help with error messages.  We use G for Tvars and E for
Evars, followed by the kind and then a number.  So, you'll
see a message like this:

  switch on type int, but case expects type `GA34 *`ETR128

This is a bit more readable but gives us all the information
we might need.

* Got rid of an extra space that was inserted for pointer types
when pretty printing.

* Added flag --manyerrors which causes the compiler to keep
going to generate all of the errors it can.

* We were allowing + e (unary plus) for any expression e regardless
of type because the parser was dropping the +.

* Avoided wrapping lines for some error messages.

* Introduced new exception Tcutil::AbortTypeCheckingFunction
which I use to back out of type-checking a function so we can
better recover without the compiler falling over.

* Tried to improve various error messages.  For instance, I
tried to consistently use "member" instead of "field".

* Bug in scope2string.

* Introduced exception Env_error which is thrown when we
end up trying to get the function-part of an environment,
but we're at the top-level.  This gets caught in various
places to avoid crashing the compiler.

* We were allowing - e (unary negation) for non-numeric types.

* We were missing a few cases where we should check that
we're not making a copy of a unique pointer.

* We don't fail in is_noalias_pointer_or_aggr when encountering
an unexpected type.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2684 581db07f-6e62-4d40-9733-d432c4eebb07