Added a tests/fail2.out and the appropriate lines to the failtest target.
Generated a handful of test cases before deciding this was a task
inappropriate for sunny summer afternoons.  I'll work on it soon, when
it's dark out.

Also, I'm not really looking at the code for now because "coverage"
for the flow analysis has more to do with the code being checked than
the code doing the checking.  (There are a lot of different ways to
get "dereference of possibly uninitialized pointer" for example.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2691 581db07f-6e62-4d40-9733-d432c4eebb07