Changed the syntax to allow a semicolon after a (non-nested) function definition. Rationale: GCC accepts this for C code (and so do many other C compilers). Also, it just looks better when you define a function using a macro:
#define DEFSOMEFUN(fun) void fun() { /* do some stuff */ }

DEFSOMEFUN(foo);

just looks better than just

DEFSOMEFUN(foo)

because people (me! :) ) instinctively end lines like that in
a ';'. This was what triggered it for me, but the GCC compatibility
argument is the most compelling argument I guess.

(BTW, I added this only for non-nested function declarations; in all
other situations a ';' after a compound statement will probably
be accepted as an empty statement.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2695 581db07f-6e62-4d40-9733-d432c4eebb07