* Added "test_src_compiler" target that tests if the $(BUILDDIR)/cyclone   compiler can compile itself and that this results in the same binary   (as suggested by Trevor). * Safeupdate now requires test_src_compiler.
The downside of this addition is that safeupdate now doesn't work if
the compiler's code generation changes significantly (so that the
binary output changes). Maybe add a target safeupdate_noselfcompile
that you can call if you want to bypass this test? => cheap trick
is simply to touch test_src_compiler, as I haven't made it a .PHONY
target.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2702 581db07f-6e62-4d40-9733-d432c4eebb07