Fixed bug that prevented us from taking the address of a struct member of abstract type (even if it's been instantiated.)  For instance, suppose you have:
struct Foo<`a::A> { int x; `a other; };

and suppose you have a variable p that is a pointer to a
struct Foo.  We would allow in the type-checker for you
to do &p->other, but the code generator would not allow
this.  The problem is that if `a is instantiated with
double, then the offset of other is one thing, and if it's
instantiated with char or int or something else, it could
be somewhere else.

I've fixed this by forcing the other field to have the
attribute __aligned__ which says that GCC should put
the field on the maximum boundary for the given architecture
(usually a multiple of 8).  In addition, any instantiation
gets this attribute.  That way, you can get at the address
of the field even if the type is abstract.

This allows us to do things like put a datatype at the
end of a struct without boxing it as in:

  datatype Bar { Int(int); Double(double); ... }

  typedef Foo<datatype Bar>@ foo_bar_t;

which is what Matthew needed in the scheme code.  I checked
that the scheme code now compiles.

Some of the other generated files changed since we use
these kind of structs in the abstract syntax (i.e., the
generated C code now has __aligned__ attributes in it.)

Although this is GCC-specific, it's at least portable...


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2879 581db07f-6e62-4d40-9733-d432c4eebb07