Added a new tool, errorgen inspired by C.L. Jeffrey's recent TOPLAS 25(5) article.  The tool is intended to help us print out decent parse error messages.
The point of all of this is that we can get very good error messages
for certain situations without having to do any work except generate
example syntax errors with an associated error message.  

The tool takes as input a file filled with Cyclone fragments and
separated by /* comments */.  (See src/errorgen_input.cyc for an
example.)  The fragments are intended to have syntax errors.  The tool
runs the compiler on each fragment with a special flag (--parsestate)
which causes the compiler to print out the parser state and token
values.  Errorgen then spits out these values, together with the
comment associated with the code.  (See src/parse_errors_gen.h).  I
put all of this information into a static table which is used as
follows:

The parser generator has been modified so that yyerror takes, in
addition to a string, the state and the token as arguements.  For the
compiler, I've moved yyerror into a new file parse_errors.cyc.  This
file #include's parse_errors_gen.h which was generated from the
errorgen tool.  The new version of yyerror now checks to see if
there's an entry in the table corresponding to the state and token in
play at the time of the parse error.  If so, it prints out the
associated error message.

Note that if the grammar changes, then the states (and tokens) might
change for the associated error message.  Thus, we really should
modify the bootstrapping so that if src/parse.y or
src/errorgen_input.cyc change, then we re-run errorgen with the new
compiler on src/errorgen_input.cyc to generate a new
parse_errors_gen.h.  Then we should rebuild the compiler.  I'm too
lazy to figure out how to do this and will leave that to you Makefile
wizards.  For now, we can do this manually.

From now on, if you get just "parse error" as a message, you 
should figure out what's going wrong (e.g., missing ';') and
cut the code down to a small example, then stick the fragment
in src/errorgen_input.cyc along with the message in a comment.

TODO:  The approach needs to be rethought to handle situations
where the next token doesn't matter, so that we don't have
to have a table entry for all of the state/token pairs.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2885 581db07f-6e62-4d40-9733-d432c4eebb07