Rewrote the allocation routines in runtime_cyc.c to be a little faster, especially for the case where you're allocating in a region.
Added support for rmalloc_inline(r,e).  This calls a version of
rmalloc that should be inlined and work relatively fast for non-heap,
non-rc, non-unique allocation.  Didn't bother to add support for
rnew_inline or rcalloc_inline.  I'm not sure it's worth it.

In addition, when compiling the output of the compiler, you can pass
the flag -DCYC_NOALIGN and this will cause us to skip the padding that
is usually added for alignment.  This flag should only be used on
architectures where alignment isn't necessary.

Added a new type constructor BuiltinType so that we can easily add
stuff like __builtin_va_list that GCC provides as an abstract type.
You have to specify the kind, which can be kind of tricky to figure
out...

Moved a type-related function from tcexp to tcutil.

Fixed bug in rewrite.txt involving tabs (thanks Yanling.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2941 581db07f-6e62-4d40-9733-d432c4eebb07