Changed config/buildspecs so that we do not have -elim-statement-expressions on by default.  This avoids possible short-comings in the tovc pass w.r.t. control flow in a statement-expression.  The problem is that we lift some statements that can't be turned into expressions into separate functions.  If the control flow ends up spanning those functions, then we're hosed.
Added warnings when we run into a continue, break, goto, or return
within a statement that gets hoisted out to a function as part of the
tovc pass.  Right now, the warning does not get tiggered when
bootstrapping the compiler, which just means we're lucky (see above).

The options in cyclone.cyc are now statically allocated (some macro
hackery that someone else might be able to improve.)  I added two new
options, -help and -usage which print out the available options.  (To
tie that knot, I needed to statically allocate the options.)  It would
be extremely useful if we supported &<initializer> within
initializers, and took this as short-hand for creating a variable x
initialized to <initializer>, and then replace that with &x.
For instance:

options_t options[] = { &$(...), &$(...), ... };


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2974 581db07f-6e62-4d40-9733-d432c4eebb07