Statements no longer have .non_local_preds (nobody was using it, so jump-analysis no longer produces a table of predessors, though doing so is easy and just commented out)
Break_s, Continue_s, and Goto_s no longer carry statement-options.
The control-flow analysis uses the succ_table from the jump-analysis.

Bug-fix in jump analysis to set "next statement" in loops correctly.

Started ripping stuff out of tcenv.cyc.  There's probably stuff left
in there that's a no-op now and should be cleaned up.  Also need to
think more about how to compute fallthru-destinations in a good way.
(The type-checker needs to know the destination, so we still do this
the "old way" of mutating the abstract-syntax to point to the destination.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3007 581db07f-6e62-4d40-9733-d432c4eebb07