In preparation for moving name-resolution to an earlier pass, changed Var_e so that it doesn't carry its own qvar_t, except in the case that the var is not yet resolved.  This avoids maintaining a weird invariant that a Var_e carries a qvar_t that is the same as the qvar_t in the binding it points to.  Now we just have one.
The only glitch in making the change was in toc.cyc where we
occasionally modify a vardecl_t's qvar_t.  If we do that after we
added the old qvar_t to the varmap we're screwed.  So I had to change
a few lines to add the new qvar_t to the varmap.

Although I didn't try it, it seems to me this change lets us get rid
of the varmap in toc.cyc altogether.  The only reason it's actually
used is in pattern-compilation and it would now suffice to mutate the
name field of the appropriate vardecl_t.  I encourage someone to try
this. :)

Bootstrapping seems to take the same amount of time, or perhaps speed
up by 1 second out of 75.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3011 581db07f-6e62-4d40-9733-d432c4eebb07