Re-run configure (added a new file).
Added an extra pass that forces the evaluation order of the generated code.
In particular, the order is left to right in general, but right to left for
assignments.  This allows the generated code to be a lot simpler, since
lvalues are preserved (see below).  To skip this pass, you can call
the compiler with the option -no-seq-c.

The basic strategy takes some expression e consisting of subexpressions
e1 ... en, and replaces e with a statement expression which assigns the
results of e1 ... en to temporary variables and replaces the occurrence of
the original expressions in e with those variables.  I had to make a bunch
of small changes to toc.cyc to make sure that types in expressions are
propagated correctly so that we can declare the types of the temporary
variables.

There are two basic optimizations that avoid creating temporary variables
unnecessarily (i.e. leave the original subexpressions alone when possible):

1) The final subexpression en does not need to be moved.  This makes
   generating code for assignment expressions simpler when the order is
   right to left, since the left-hand-side then need not be stored in a
   temporary (which would require we take its address to assign to it).

2) Avoid moving based on a simple side-effect analysis.  In particular, if
   only one expression contains a side-effect and the rest are constants,
   there is no need to rewrite the expression e.  Also, read-only
   expressions (like variables and constants) can coexist properly (i.e.,
   none need be rewritten) alongside a function call that has effectful
   control flow (i.e. could throw an exception), but is otherwise read-only.
   This is not true when two subexpressions could throw exceptions, since
   the order then matters.

These optimizations are pretty effective.  The size of the genfiles did not
grow appreciably, and the compile time for the entire system is pretty
similar (2m42s as opposed to 2m28s on my Linux laptop).  The next step is
throw away lots of stuff in the flow analysis ...

Also fixed a bug in the flow analysis that failed to correctly eliminate a
null check.  Made some slight adjustments to permit compilation on BSD.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3072 581db07f-6e62-4d40-9733-d432c4eebb07