Added support for "anonymous nodes" as Greg and Dan have referred to them, in preparation of separating out the alias analysis from other parts of the flow analysis.
The insight is that a linked datastructure consisting of unique pointers
might as well be thought of as flattened, since there is only one path to
each element.  Thus, we could extend our notion of "place" in the flow
analysis to not just be a root and a list of field offsets, but rather a
root and a list of both field offsets and unique dereferences.  For example,
given

struct foo {
  struct bar {
    int *`U x;
    int *`U y;
  } *`U b;
  int *`U z;
};
struct foo *`U f = ...

Then we would represent f->b->y as a place (f,[*; .1; *; .2]).  In the
current arrangement, we create fake MallocPts for intermediate pointers, so
f->b->y would be represented by a series of places, e.g. in the flow
dictionary we'd have something like:

f              ==> AddrOf(MallocPt(FOO))
MallocPt(FOO)  ==> Aggregate(AddressOf(MallocPt(BAR)),UnknownAll)
MallocPt(BAR)  ==> Aggregate(UnknownAll,UnknownAll)

In this (the old) arrangement, we would occasionally have to generate fresh
MallocPts to give names to unique locations.  In this case we have two
problems: (1) the names are chosen somewhat arbitrarily, which can have the
effect of being overly conservative on joins, and (2) the MallocPts
themselves are totally unnecessary, since they are really needed to
represent aliasing, but of course there isn't any for unique pointers.  In
the new arrangement, we add a new absRval called UniquePtr(rval) to
represent a unique pointer, in lieu of AddressOf(MallocPt(...)) that we have
now.  This corresponds to a * in the place path shown above.

In the code, I basically had to change allocation code to not generate
MallocPts for unique pointers, and I changed anal_derefL to construct a
place based on the existing unique path, not the MallocPt being derefed.  I
also cleaned up the code for updating the relations for this function.

When I started putting this code in, I was convinced it was the right thing
to do.  Now I'm less sure, since the result is a bit more code, another case
to handle for analyzing dereferences, and possibly less potential for
modeling local aliases to unique pointers.  OTOH, joins should work out
better, and this style more clearly matches our intuition of a unique
pointer being the capability for the tree that it transitively points to.
We'll see.

Other changes:

(1) Added a special expression form __cyclone_pragma__(p).  This is useful
    for inserting notes to the compiler.  I use it for debugging the flow
    analysis.  In particular, you can do:

    void foo() {
      int x;
      __cyclone_pragma__(print_flow);
      x = 2;
      __cyclone_pragma__(print_flow);
    }

    This will print out the flow dictionary at each of the two points.
    Because it's an expression form, you can put it in sequence expressions,
    etc.  In terms of code generation, this directive is identical to the
    constant 0.  The other directives I added were "debug_on" and
    "debug_off" to help selectively print flow debugging messages when the
    DEBUG_FLOW #define is turned on.

(2) Killed the notion of "This" initialization.  That is, we now just have
    All of None.  I found no ill-effects on the Cyclone code I tested
    (the compiler, the tests, and MediaNet) after making a tweak to the code
    that uses the initializes() attribute.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3085 581db07f-6e62-4d40-9733-d432c4eebb07