* Fixed bug that Dan found regarding region ordering information for opened dynamic regions.  In particular, if we open dynamic region `r, then none of the current regions in scope are assumed to outlive `r (it maps to the empty effect in the PO map.)  However, for any lexical region `r2 created within the scope of an opened region `r, we know that `r outlives `r2.
Mike, one thing that confused me is Tcpat::check_alias_coercion().
We're supposed to check whether the old type is a sub-type of the new
type, under the assumption that the old region being aliased is a
sub-type of the freshly introduced region (e.g., you can promote from
`U to the new alias region `r.)  The code actually adds the assumption
`U <= `r which doesn't seem right.  In particular, if we alias a
pointer into `RC or `x::TR or `y::R then we should be adding the
assumption that `RC <= `r, `x <= `r, or `y <= `r respectively, and not
`U <= `r.  But that got me to wondering why the code in string.cyc
that does aliases with pointers into a TR actually type-checks.  I
suppose we're assuming that the TR region outlives the region
introduced by the alias stuff naturally, and so it falls out.

We still need to rule out deep aliasing for `RC.  Should we also
rule out deep aliasing for `r::TR?  And if so, what should we do
about the code in string.cyc that aliases `r::TR?


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3095 581db07f-6e62-4d40-9733-d432c4eebb07