Run configure.
Changed the way that cyclone and buildlib handle the -b flag.
Previously they just passed it to gcc, however, gcc's conventions
for cross-compiling are not consistent.  For example,
gcc -b arm-elf will invoke an executable arm-elf-gcc-XX,
where XX is a version; but typically the arm-elf cross-
compiler is installed as arm-elf, without a version number.
So, we no longer pass -b to gcc, we just use it to find the
appropriate cycspecs file, which tells cyclone how to invoke
the right gcc.  The cycspecs file can use -b if desired.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3144 581db07f-6e62-4d40-9733-d432c4eebb07