A little hacking while at a PI meeting:
* Fixed bug in binding.cyc (not really a bug, but Dan wasn't sure what to
do here) where we had a declaration of a variable whose type included an
anonymous enum.  Before, the tags weren't getting entered into the 
environment.  Now they are entered into the current namespace.  I think
we really ought to get rid of anonymous enums and just insert a 
freshly generated name.

* Added callgraph.cyc and callgraph.h which can be used to compute
(surprise) a callgraph for a set of declarations, and also the
set of strongly connected-components.  The goal is to use this
to support region inference for mutually-recursive functions in
Tempest (and possibly for Cyclone functions declared static.)

* Fixed bug in documentation regarding syntax for explicit region
orderings.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3406 581db07f-6e62-4d40-9733-d432c4eebb07