Fixed bugs that prevented tests/require and tests/streambuff from going through.  I'm not exactly sure what happened here, but a test of the form  (i == 1)  where i has type tag_t<`i> was not being tracked by the relations.  The reason was that the type-checker was inserting a coercion to int so that we were testing ((int)i == 1) and the flow analysis doesn't bother to look at relations on signed values.
So I changed a couple of things.  First, the relations now
look at equality and inequality for any integral value since
there's no difference between the signed and unsigned tests.
This covers the particular problem here, but not when we have
an inequality involving a tag_t.  So second, I changed the type-
checker so that instead of promoting a tag_t to an int, it
gets promoted to an unsigned int.  That broke some of the
code in evexp which was not expecting to run into a tag_t
variable.  So, I fixed that as well.  

Finally, I commented out the include of stdlib.h in ap.cyc
in an attempt to minimize the dependencies on arch-specific
includes.  We still depend upon limits.h for LONG_MIN and
LONG_MAX which I think we can hack around.

There's still a lot of null checks that shouldn't be in
there, according to the warnings we're generating.  It
seems to have something to do with the fact that when we
promote an array A to a fat-pointer, we lose the fact that
it's definitely not-null.  I will look into this and try
to fix it as well.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3414 581db07f-6e62-4d40-9733-d432c4eebb07