Rerun configure.
Extended extern "C include" to include a cyclone_override { decls }
component.  This can be used to provide Cyclone definitions for
prototypes appearing in the C-included part.  This is analogous to the
functionality I added to buildlib a while back, but I've expanded it
to variable definitions, not just type definitions, and it now allows
more Cyclone types including polymoprhic ones.  An example can be
found in tests/cinclude.cyc.

There appears to be a bug in the callgraph generation for the last
check-in; it barfs on scheme.cyc in the tests/ directory.  So I
commented out the calls to Callgraph in tc.cyc until it gets fixed.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3427 581db07f-6e62-4d40-9733-d432c4eebb07