Cleaned up absyn.cyc.  Only substantive difference to other files is that many globals have become constant functions, e.g., exn_type became exn_type().
Then I noticed that two tiny regression tests in fail1.cyc were
(not) failing and I spent the rest of the evening fighting attributes:

(1) We were not detecting that packed and aligned are illegal attributes
for function definitions.

(2) For reasons I don't understand, moving this check before we call
    check_fndecl_valid_type fixes that problem.

(3) But then I also noticed that cyctest was core-dumping because we
    dropped an important __attribute__((stdcall)) from a function
    definition.  I have no idea when this started happening -- the
    likeliest is with my earlier check-in that changed struct Fndecl, but
    after spending a lot of time examining the cvs diffs, I don't see
    how.

(4) The problem has to do with how we split the attributes up between
    the attributes field in the struct Fndecl and the attributes field
    in its cached_type.  We currently _partition_ the attributes, with
    those affecting the type going in the cached_type attributes.

    (a) If the struct Fndecl attribute field has _all_ the attributes,
    then Tc::merge_binding fails because the attributes for a
    prototype and a definition no longer match up.

    (b) You can fix (a) by having prototypes also have _all_ the
    attributes in the of struct Vardecl, but for some reason that
    didn't work either. [Sorry, it's late]

    (c) The attributes in the cached_type were being ignored by the
    printers.  So now I look at the cached_type and move its
    attributes back to the struct Fndecl when printing a Fn_d.

Since this solution is clearly not what we were doing before, I wonder
what we _were_ doing before.  If it was a variant of (4b), it sure
doesn't look to me like it's been that way for a while -- I'm curious
when the problem was introduced, but probably not curious enough to
find the answer.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3428 581db07f-6e62-4d40-9733-d432c4eebb07