Main thing I did was get parser-generation to produce less code, especially C code:  * instead of each case declaring a local thin-pointer yyyyvsp, each  case just sets one defined in the same scope as yyvs.  The bound for  yyyyvsp is the maximum length of any rule.  To ensure the cast to a  thin pointer never fails, the check for resizing yyvs subtracts this  maximum length in the necessary way.  * a static function yyget_yyyyvsp does the cast to a thing pointer.
The upshot is: yyparse is smaller because each case just makes a call
to yyget_yyyyvsp.  This is especially true at the C level because
yyget_yyyyvsp is now the centralized place for doing a check_null, an
untag_fat_ptr, and a _fat_ptr_plus, which are each macros that were
being expanded 538 times in the same function in parse_tab.c.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3455 581db07f-6e62-4d40-9733-d432c4eebb07