Added constraint solver to vcgen.  Still need to handle the signed/ unsigned distinction and a few other things before this is ready for prime time.
The constraint solver has the unfortunate habit of blowing up the
assertions (it effectively has to calculate the DNF).  Right now,
I have a variable (max_paths) that can be used to throttle the
search.  Right now, it's set to 33 distinct paths which means that
if you end up with more than this number of paths in a function,
the prover falls back on syntactic matching instead of constraint
solving.  Obviously, you can crank up the number of paths with
a tradeoff in both space and time.  I've worked hard to minimize
the overheads while keeping the code reasonably clean.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3518 581db07f-6e62-4d40-9733-d432c4eebb07