Added "relative region" feature to Cyclone.  The goal is to avoid having to pass around explicit region handles all the time.  This is an intermediate step toward implementing autorelease using an implicit region stack, as in MacOS.
You need to rerun configure.

Here's how it works.  I introduced a special region variable `C which
represents the topmost lexical region.  If no topmost lexical region
exists, it refers to the heap.  Then, you can define routines that
always allocate into the topmost region.  E.g.

int *`C alloc_int(int x) {
  return rnew (Core::current_handle()) x;
}

The current_handle() function acquires the topmost region handle from
the region stack and returns it.  Thus, I can call this function in
different contexts, and have it allocate in different regions, without
having to explicitly pass a region handle.  We also get all of the
outlives information we expect:

int foo() {
  int *`C p = alloc_int(1); // `C == ? (`H)
  region r;
  int *`C q = alloc_int(2); // `C == `r
  region r2;
  int *`C r = alloc_int(3); // `C == `r2
  int *`C s = q;            // OK, `r1 > `r2
  q = p;                    // OK, `C > `r
  // p = q;                 // disallowed
  // q = r;                 // disallowed
}

This mechanism thus makes it easier to convert existing C code that
uses malloc to use regions instead, as we can simply insert region
declarations and a few annotations, and we don't have to pass an extra
parameter around.  However, it's still not as great as we would like,
because we do need to add `C annotations for functions in which we
perform allocation.

`C is implemented as a special type variable.  You will be prevented
from trying to refer to `C in a type list (e.g., no typedef int *`C
c_int<`C>) and other places.  Just after the binding phase, all
occurrences of `C in function bodies are substituted with the actual
topmost region.  This is what prevents the disallowed cases in the
above example.  Occurrences of `C in function prototypes are left
alone, as are occurrences within function bodies that are prior to a
region declaration.  Thus, after this phase, the above code looks as
follows:

int foo() {
  int *`C p = alloc_int(1);
  region r;
  int *`r q = alloc_int(2);
  region r2;
  int *`r2 r = alloc_int(3);
  int *`r2 s = q;           // OK, `r1 > `r2
  q = p;                    // OK, `C > `r
  // p = q;                 // disallowed
  // q = r;                 // disallowed
}

During typechecking, we always add `C to the capability and
environment of functions while checking them.  This is why the first
alloc_int() above is OK.  Whenever we call a function that refers to
`C in its prototype, we can instantiate `C only with the topmost
region.  This is inferred automatically in the above example, but if
the programmer were to try to instantiate `C manually, anything other
than the topmost region would be disallowed.

This required remarkably few changes to the main part of the
compiler.  The big change was the initial substitution pass, added as
a separate phase.

There is an example in tests/test_currgn.cyc.

Also, added stage timers to cyclone.cyc so that when you run in
verbose mode, it prints the time spent in each phase.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3527 581db07f-6e62-4d40-9733-d432c4eebb07