* Changed control flow checking of try blocks
  void foo(int *`U a) {
    int val;
    try {
	let alias<`r> int *`r b = a;
	...
    }catch {
       default: val = *a;
    }
  }

This was previously disallowed since a would be treated as consumed
in the catch block, and in the try's continuation.

Modified the represenatation of the "Consumed" abstract value so that
we now keep track of how the value was consumed. In case it was
consumed by a "let alias" construct we don't consume the value in the
try flow, although it is consumed within the scope of let.

The right way to do this would probably be to integrate it with the
jump_analysis so that we unconsume before checking the catch
block. But, the way the recursion is structured right now makes it a
pain to keep track of the pattern variable depth inside the try
block. Should revisit this at some point to make it cleaner.

* Added an option "--cifc-inst-tvar"

This controls the behaviour of cifc introduction of
type-variables. When the flag is set cifc prefers to instantiate type
variables with values of appropriate kind (void* for B, `H for R etc.)
intead of propagating the polymorphic parameters through the type
hierarchy.

* Updated the no_throw attribute checking to allow some simple
expressions in a function body to not be embedded inside a
catchall handler.

* Fixed some bugs

Jump_analysis.cyc computation of pop-tables was causing the code
generator to pop the runtime stack even when exiting the scope of a
dynamic region.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3626 581db07f-6e62-4d40-9733-d432c4eebb07