* Fixed bug in vcgen for checking @requires clause -- had the wrong location.
* Added a debug flag to support more verbose printing of assertion info 

* Modified string.h and string.cyc to take advantage of @assert's.  I'll
be doing more of this in the future as we get the constraint solver
whipped into shape.  Right now, all we have are three functions that
use the static assertions.  strlen(s) now @ensures that the result is 
no bigger than numelts(s).  umin(x,y) is a useful inlined 
function that @ensures its output is no bigger than its inputs.  And 
strcmp(s1,s2) now has no dynamic asserts and no bounds checks.  I started
to modify ncmp to do the same, but alas, we need support for congruence
closure to verify the necessary assertions.  So we're going to add that
next...


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3778 581db07f-6e62-4d40-9733-d432c4eebb07