* Got rid of equivalence classes -- use the graph constraints instead. Though this results in bigger graphs, it's more general and simplifies the code.
* Keep graph rows sorted to make lookups faster and to help with
congruence closure (see below).

* Added explicit pass for doing the congruence closure.  In essence,
we look at every pair of compound terms F(t1,t2) and F(u1,u2).  If
the graph says that t1 == u1 and t2 == u2, then we equate the two
F terms in the graph.  I'm currently invoking this closure only when 
we process a disjunction (or at the end of constraint solving).  That
helps to amortize the closing that we do.  Might be good to add a
"dirty" bit to graphs to record whether they've been closed, to
avoid re-closing when we have nested-or's.  

* A couple of changes to help with building on FreeBSD (not there yet
though.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3820 581db07f-6e62-4d40-9733-d432c4eebb07