* Fixed printing of @aqual stuff. Basically, @aqual(ALIASABLE) should never appear in an error message. Also fixed a problem with the pretty printing of qualifier bounds.
* Added a rule in the typechecker that disallows qualifier bounds from
appearing in locals etc. Take the following example:

void foo(`a\A arg) {
  `a\U hd = arg;
}

The bound on hd below was previously ignored and this function would
typecheck. This is now an error -- qualifier bounds may only appear in
function prototypes and in aggregate fields.

* Moved definition of struct _RegionPage to cyc_include.h. It used to
be defined in either runtime_memory.c or cyc_include.h depending on
CYC_REGION_PROFILE. Hence the changes to all the genfiles.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3947 581db07f-6e62-4d40-9733-d432c4eebb07