Added the extended occurs check to effect unification
Unifying effect `e1 with `e2 now proceeds as follows:

1. If `e1 does not appear in `e2 and if every type variable that
appears in `e2 is also in the scope of `e1
(i.e. Unify::occurs(`e1,`e1->tvars, `e2) passes) then `e1 is unified
with all of `e2.)

We need this so that unification can handle the following:

int *`r+`H a = NULL;
int *b = a;

2. If the occurs check fails and `H appears in `e2 then `e1 is unified
with `H. Note that we need to check that `H appears in `e2 since `H
no longer appears in every effect.

Previously when checking that `r was in the effect {`r1, `r2}, `r
might have been unified with either `r1 or `r2 depending on which one
passed the occurs check. If neither passed the check then `r would be
unified with `H.

Not mimicking the previous behavior is useful since it allows for
some annotations to be eliminated. For instance, the two annotations
in tcutil.cyc and pratt_prover.cyc that I reported previously are no
longer necessary. No additional annotations were required either.

For example:

void foo() {
  int *a=NULL;
  f(a); //a becomes int *`H since `H is in the capability
}

3. If the first two conditions do not apply and then unify `e1 with
the first component `r of `e2 such that occurs(`e1, `r) passes and
unify `e1 with `r.

This case is just to make things roughly equivalent to the previous
behavior described above. This rule was never actually used for any
code in the src, tools, lib and tests.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4159 581db07f-6e62-4d40-9733-d432c4eebb07