Many bug fixes for vcgen.
* Fixed a bug for New_e.  We were treating New_e of an array or
comprehension badly --- as if it were a pointer to one object
which held an array.  

* Fixed a bug in vcgen regarding comprehensions.  A comprehension of
the form { for i < 256 : 0 } is considered "constant" by Evexp, but
we really want vcgen to "run" this code.  More generally, it may be
better to just eliminate the "is_constant" hack at the top of vcgen_rexp
to avoid problems like this and let vcgen figure out whether or not
things are constant.

* Fixed the analysis for comprehensions.  This was very broken
and ended up returning a context that was inconsistent (i.e.,
false) almost every time.  It would be very good to factor out
all of the vcgen code that deals with finding loop invariants
so that we avoid this sort of problem.

* Fixed a bug in vcgen regarding declarations.  We must treat a
declaration "int x = e;" by first assigning a logic variable to
x (representing that x is initialized by the machine with some
random value) and then interpret e.  This is really only a problem
with "bits-only" things since they need not be initialized before
they are used.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4255 581db07f-6e62-4d40-9733-d432c4eebb07