Created a new datatype for grammars, consolidates rules with auxiliary information such as FIRST, FOLLOW.
Added a function to determine the LAST set of rules.

Added the ability to check for right recursive symbols.

Added a check for white edges (a symbol has a white edge if it can
start or end with whitespace). TODO: figure out if we are just going
with spaces and whether left/right distinction is important.

Instead of HT use HTAB in prose elimination; that's what RFC 2234
uses.

Prose replacement bug fixes for token, field-content.

Added a new ykbuf that takes input from a parser.

Disabled ws_definitions, soon to be removed.

Made minus_elim and prose_elim imperative.

Overhauled character set annotations, which are now imperative; added
a field to rules to cache the character set status (css).


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4267 581db07f-6e62-4d40-9733-d432c4eebb07