Some more improvements to vcgen.  The coolest one is at the bottom:
* Got rid of redundant closure when we have nested Or(..) nodes
in constraint solver.  This significantly sped up compilation of
the parser.  

* Two new flags for controlling vcgen resources:  --max_vc_paths <n>
determines the maximum number of paths that the constraint solver
will explore before giving up (default currently 17.)  Cranking
this up may result in fewer checks, but will mostly just slow
compile times.  The other flag is --max_vc_summary <n> which controls
the maximum size of a summary assertion calculated for procedures
without weakening (default currently 500).  Cranking this up will
provide more detailed inter-procedural information to be carried 
across function calls (see below).  It would be cool to do a search
to figure out what a good sweet spot setting is for these knobs
that minimizes compile time while maximizing theorem proving.

* More bug fixing in vcgen_rexp -- we're treating some values as
constant expressions (e.g., &x when x is a global) which we should
really process as terms.  I'm thinking of getting rid of the whole
if (is_const_exp(...)) except that I need evexp to evaluate 
constant expressions.  Someday, I'm going to have to rework this.

* Fixed bug in vcgen for function calls.  We weren't updating the
exception handler pre-condition unless there was a @throws clause...

* Added "invocation" terms to Addr(...) and Alloc(...) terms.  The
invocation is used to distinguish distinct invocations of a procedure,
which I needed to ensure soundness of the stuff below.  It essentially
is used to avoid treat two allocation terms that conceptually arose
from two different invocations of the same procedure from being
confused as equal.  

* Major coolness:  I now record a calculated strongest post condition
for functions that gets used in subsequent calls to that function, 
which has the effect of "inlining" the function calls as far as 
vcgen is concerned.  This extends vcgen's reach to be inter-procedural
without having to do much.  In essence, within a module, any procedure
that is defined before will have its post-condition (possibly weakened)
available for subsequent callers.  So for instance, if you have:

int f(int x,int y) { return x+y; };

void g() { int i = f(3,4); int j = f(5,6); }

Then vcgen will figure out that i is 7 and j is 11.  Just as imporantly,
it will know that memory is unchanged across the calls.  Effectively,
I'm summarizing the effect of a procedure call as a function:

  fn t1,...,tn,mem -> Exists v1,....,vm.A

At a call site, I apply the meta-function to the actual term arguments
and memory to produce a context-sensitive post-condition.  

A few notes

(1) I have to be careful when "opening" the implicit existential
to alpha-vary the logic variables.  By adding an explicit variable 
to represent the invocations for the Addr(...) and Alloc(...) terms, 
I'm assured that these things aren't confused as equal across two 
different calls through alpha-conversion.  (This makes me think
that we should worry about these issues in @requires/@ensures/@throws
clauses -- i.e., the conversion from expressions to assertions
could generate free variables which we must bind and alpha-vary
appropriately.)  

(2) I have to throttle how big the assertion can get.  Ironically,
when trying this out on vcgen.cyc, we run out of memory, but every
other file goes through without throttling.  The reason it chokes
is that it summarizes a function like vcgen_test using an assertion
with over 11,000 free logic variables (in turn because vcgen_test
makes many calls to vcgen_rexp which is huge.)  Lord knows how big the
assertion itself is!  Anyway, vcgen_test gets called many times within
vcgen_stmt, each time demanding that the assertion be alpha-varied.
So vcgen_stmt's assertion becomes absolutely enoromous and we run
out of memory.  

So I've arbitrarily limited the assertions that I'll record to have at
most 500 free variables in them.  If the assertion is bigger than this,
then I weaken the assertion by only pulling out the common conjuncts.
That should still be pretty good.

(3) it doesn't always help!  The reason is that the constraint prover
gives up when it hits an assertion with a large number of disjunctions
in it, to avoid an exponential blow up.  When we effectivley inline
the summary for functions, we carry those disjunctions over (instead
of just summarizing everything with one variable) so the prover is
more likely to give up.  

As a result, the results were quite mixed.  A few checks weren't
eliminated in some files, and in a very few files (notably lib/ap.cyc)
a ton of checks were eliminated.  The reason is that ap.cyc uses *
where it should really use @, and the inlining picks up the fact that
a function which does nothing but allocate something returns a
non-null pointer.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4299 581db07f-6e62-4d40-9733-d432c4eebb07