Changed the fundamental strategy for the converting an assertion to a set of constraint graphs.  Before, we were in essence doing the expansion in a breadth-first fashion.  Now, I do the expansion in a depth-first fashion so that if we expand out an entire path and cannot prove it's inconsistent, we give up right away.  This turns out to be significantly faster than the old approach, particularly in the cases where we can't prove something.  (With the old strategy, we would've also explored the other paths.)  The code for doing this is quite a bit simpler, too.  I'm not sure why I didn't think of doing it this way the first time...
To give a feel for the improvement in speed, on my laptop, the old
compiler took 3m47s to bootstrap and with the changes I checked in, it
took 2m50s.  Part of the improvement is that I was able to just turn
the conversion into a big loop by just having explicit success and
failure contexts represented as explicit stacks.  (What's fun is that
I originally wrote the code in ML as a sort of CPS-based function,
then manually closure-converted and defunctionalized it, and ended up
with the pretty code that's in pratt_prover::inconsist right now.)

This was fast enough that I was able to squeeze in one more nice
change without adversely affecting compile times: Before, when we
encountered too many paths, we just gave up.  Now, we simply widen the
two paths into one and continue.  This allows us to get rid of quite a
few more checks throughout the compiler.  Doing this slowed
bootstrapping down to 2m57s, so it seems worth it for the extra
precision.  

Another nice thing about this change is that we can delay doing
any of the closure passes until we've constructed the whole graph
for a given path.  That ensures that we don't end up re-calculating
expensive closure operations (e.g., congruences, shortest-paths,
etc.)  

Next up:  Yanling's suggestion to break non-integral terms out
of the graphs seems promising.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4563 581db07f-6e62-4d40-9733-d432c4eebb07