* Bug fix in tcexp:  we had "if (... && !allow_valueof)" instead of "if (... && !te->allow_valueof)".  This compiled because allow_valueof was a function name that was in scope.  So !allow_valueof always evaluated to false.
* Added extra reduction to the update() smart constructor for terms.
If we have update(m,offsetf(a,i),v) then this reduces to 
update(m,a,aggr_update(select(m,a),i,v)).  This is effectively
an eta-expansion which allows more of the updates on memory
to collapse.  

* Added extra reduction to the aggr_update() smart constructor for terms.
In particular, we now sort the updates by index.

* Modified merging of assertion maps to do deeper unification for
terms.  In particular, this does a form of eta-expansion for update
and aggr_update terms that allows more of them to merge.  Unfortunately,
we still lose all information about memory when we abstract for 
loops, so we still need to do some work here to realize any benefits.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4648 581db07f-6e62-4d40-9733-d432c4eebb07