* improved vcgen a bit:  when x is a (thin) pointer, we no longer treat (int)x and (unsigned)x as terms that are distinct from x. We were treating them as distinct before just to be conservative, but this interferes with checks of the form "if (!x) ..." and "@assert(x)" etc.
* improved prover a bit:  subst_close was doing simultaneous
substitution of "better" terms for other terms.  Some of these
substitutions interfered.  In particular, when we have a relation
of the form: T(..X..) == A && X == B && R(T(..X..),V), we might 
only conclude that R(A,V) holds instead of also concluding 
R(T(..B..),V).  To avoid this problem for now, I've broken the
substitution into two passes:  one that deals with substituting
for variables first, and another that deals with substituting
for non-variables second.  

* improved variable names chosen in terms so that they align better
with the user code (helps with error messages to some small degree.)

* assert(e) is now a built in operation which throws the Assert
exception upon failure.  So @assert(e) is a static assertion,
whereas assert(e) is a dynamic one.  That is, @assert(e) is 
guaranteed to have no run-time overhead and can talk about
hypothetical values (e.g., valueof(`i)) but may not be provable.
In contrast, assert(e) can only talk about computable values,
and can include assertions that we cannot prove statically, but 
in general will require a run-time check.  

* vcgen tries to eliminate assert(e)'s.  It can only do so when
e can be shown to evaluate to false at compile time, and when e
has no effects.  If the flag --warnassert is turned on (or -Wall), then 
vcgen will tell you which asserts might evaluate to false.  No warning is 
issued if the assertion will always evaluate to true, but the assertion 
includes effects.  We may want to provide an option for converting
an @assert(e) to an assert(e) when e is computable and we aren't
able to prove it at compile-time.  

* Added a @checks(e) clause to functions.  A checks clause is similar
to a @requires, except that it's performed dynamically instead of
statically.  For instance, one could write:

 void f(char ?x, char ?y) @checks(numelts(x) >= numelts(y))

The function f gets to assume the check has occurred at each call-site.
At a call-site, we insert an assert(...) corresponding to the checks
clause (but with the actual arguments substituted for the formal
parameters.)  The intention is that one should be able to decorate
libraries with liberal @checks clauses.  Clients that meet the
requirements will pay no overhead (assuming there is no effect in
the checks clause.)  Clients that meet the requirements but we can't
prove it will type-check, but will have some run-time overhead.
This is the analogue of providing implicit, checked-coercions
(e.g., from int* to int@).  

There's no real support for sub-typing on checks clauses (due to
the potential for side effects in the clause.)  However, one can
manually eta-expand to impose additional checks.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4700 581db07f-6e62-4d40-9733-d432c4eebb07